Prod-18: Add ability to view and delete suppliers.
- Create endpoint to delete suppliers by supplier_id
- Update data access layer to account for supplier_id being a int datatype
- Create view to delete and view suppliers
- Update success message view

+++ src/main/java/com/indiraactive/stockupdaterserver/dal/SupplierRepository.java
@@ -3,6 +3,6 @@ package com.indiraactive.stockupdaterserver.dal;
 import com.indiraactive.stockupdaterserver.model.Supplier;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SupplierRepository extends CrudRepository<Supplier, Long> {
+public interface SupplierRepository extends CrudRepository<Supplier, Integer> {
 
 }
